restructure chat's project folder
