+++ Sources/Chat/Types/ChatResponse.swift
@@ -1,0 +1,10 @@
+
+import Foundation
+
+struct ChatResponse: Codable {
+    let topic: String
+    let requestMethod: ChatRequest.Method
+    let requestParams: ChatRequest.Params
+    let result: JsonRpcResult
+}
+
