+++ Sources/Chat/Types/Invite.swift
@@ -1,0 +1,10 @@
+
+
+import Foundation
+
+struct Invite: Codable {
+    let pubKey: String
+    let message: String
+}
+
+
