+++ Sources/Chat/Types/RequestSubscriptionPayload.swift
@@ -1,0 +1,8 @@
+
+
+import Foundation
+
+struct RequestSubscriptionPayload: Codable {
+    let topic: String
+    let request: ChatRequest
+}
