+++ Sources/Chat/Types/Thread.swift
@@ -1,0 +1,9 @@
+
+
+import Foundation
+
+struct Thread: Codable {
+    let topic: String
+    let pubKey: String
+    //let peerName: String
+}
